refactor: migração de funcionalidades entre módulos
- Move task_publish_article de publication/tasks.py para upload/tasks.py
- Remove importação não utilizada do módulo controller
- Corrige importação de Article para usar article.models
- Remove função get_manager_info() não utilizada
- Adiciona chamada para task_check_article_availability após publicação
